pyproject.toml
[project]
name = "text-2-sql-pinterest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.8.4",
    "faker>=40.15.0",
    "google-generativeai>=0.8.6",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "mysql-connector-python>=9.6.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "sentence-transformers>=5.4.1",
    "sqlalchemy>=2.0.49",
    "torch>=2.11.0",
]
